doc: actually use the readme in the documentation
The readme was not actually in use leading to a warning by Sphinx. The
contents from the readme are relevant here so we can't just delete the
unused file, but instead fix the issue.

To better integrate the relevant contents of the readme in the
documentation and still allow the readme file to be rendered correctly
by Bitbucket, we convert it to ReST and split its content.