Remove filenames from @file tags
The Doxygen @file tag serves to identify a comment as belonging to the
file rather than any specific symbol. Often it's not even needed, but
using it explicitly may avoid some corner cases. However, a file
parameter is only needed if one wants to associate a comment with a
different file than the file that owns the comment (which might be a
questionable practice and not any of our cases).

In our cases the name of the file is redundant and poses a maintenance
burden as the files may change names and Doxygen may not know where the
comment belongs and throw a warning or, worse, silently assume it
belongs to a different file altogether.