doc: remove duplicate documentation
Each symbol should only be documented in one place or else Sphinx won't
be able to deal with references and will dutifully complain. If we want
the full index of symbols as extracted from Doxygen, we can keep those
in one single place and reference what we want everywhere else. We
replace one such case here.