doc: convert README back to markdown
Unfortunately Bitbucket does not support ReST READMEs natively, though
it seems there was at some point a plugin for it. This plugin cannot be
found anymore (?) and either way not the ideal solution. A native
solution was requested many years ago [1], but it doesn't look like it
will ever be worked on.

The options are to go back to using a Sphinx extension to import
markdown files or duplicate the relevant contents. We choose duplication
as the extension option has several downsides relative to the amount of
duplication it saves:

  - It would represent an additional dependencies for building the docs.
  - Blind include of the README.md file in the docs would limit the
    composition of the readme show on the projects landing page or
    pollute the documentation with irrelevant stuff. E.g. the title
    would be desirable in the landing page, but out of place in the
    documentation.
  - It would limit our ability to use Sphinx references and other
    features in that portion of the text, limiting the quality of the
    actual documentation.

Note this does not recover the original README.md verbatim. The table of
contents is removed (it was mostly made of missing entries) and we make
use of the better formatting from the newer doc/installation.rst.

[1] https://jira.atlassian.com/browse/BSERV-5384