mvvm/editors/longint_spinbox: implement workaround to update and propagate StepEnabled state
Normally, AbstractSpinBox internally manages StepEnabled updates.
However, in this implementation, we are not implementing/calling the
necessary methods to do so properly. As AbstractSpinBox keeps the
relevant members private, we work around this issue by triggering a
StyleChange event. This causes AbstractSpinBox to call methods that
handle StepEnabled updates for us, such as
QAbstractSpinBoxPrivate::updateEdit.